move hooks to standard providers
